apps/api/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
sqlalchemy
psycopg[binary]
python-dotenv
pydantic
email-validator
aiosmtplib
pytest
pytest-asyncio
httpx
black
ruff
mypy
